Pass unsigned value into base::span constructor in ui/gfx
base::span will shortly require that length is unsigned.

This CL adds a checked_cast for fGlyphCount in a test, as
it is not always guaranteed to be unsigned.

This is part of a larger CL https://crrev.com/c/4159822.

BUG=1385166

Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4214711
Code-Coverage: Findit <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Etienne Bergeron <etienneb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1101757}

+++ ui/gfx/render_text_unittest.cc
@@ -18,6 +18,7 @@
 #include "base/i18n/char_iterator.h"
 #include "base/logging.h"
 #include "base/memory/raw_ptr.h"
+#include "base/numerics/safe_conversions.h"
 #include "base/ranges/algorithm.h"
 #include "base/run_loop.h"
 #include "base/strings/string_piece.h"
@@ -344,8 +345,8 @@ class TestRenderTextCanvas : public SkCanvas {
     if (blob) {
       SkTextBlob::Iter::Run run;
       for (SkTextBlob::Iter it(*blob); it.next(&run);) {
-        auto run_glyphs =
-            base::span<const uint16_t>(run.fGlyphIndices, run.fGlyphCount);
+        auto run_glyphs = base::span<const uint16_t>(
+            run.fGlyphIndices, base::checked_cast<size_t>(run.fGlyphCount));
         glyphs.insert(glyphs.end(), run_glyphs.begin(), run_glyphs.end());
       }
     }
